+++ src/app/core/components/FlowStateCoreDetails.tsx
