+++ src/app/pool/components/MatchingPoolDetails.tsx
