+++ src/app/pool/components/checkout/NFTGating.tsx
