+++ src/app/pool/pool.tsx
@@ -12,9 +12,9 @@ import Dropdown from "react-bootstrap/Dropdown";
 import Spinner from "react-bootstrap/Spinner";
 import SuperfluidContextProvider from "@/context/Superfluid";
 import PoolInfo from "./components/PoolInfo";
-import Grantee from "@/components/Grantee";
-import GranteeFunding from "@/components/GranteeFunding";
-import MatchingPoolFunding from "@/components/MatchingPoolFunding";
+import Grantee from "./components/Grantee";
+import GranteeFunding from "./components/GranteeFunding";
+import MatchingPoolFunding from "./components/MatchingPoolFunding";
 import { Recipient } from "@/types/recipient";
 import { ProjectMetadata } from "@/types/project";
 import { Token } from "@/types/token";
