+++ src/app/projects/components/PoolCard.tsx
